requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
flask-sqlalchemy==3.0.5
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
orderedset==2.0.3
protobuf==3.20.3
PyMySQL==1.1.0
requests==2.31.0
SQLAlchemy==2.0.17
typing-extensions==4.7.1
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
